refactor : 메세지 클래스 단일화
