--- src/main/kotlin/racingCar/etc/ErrorMessage.kt
@@ -1,10 +1,0 @@
-package racingCar.etc
-
-/**
- * 에러관련 메세지를 모아둔 클래스
- * Created by Jaesungchi on 2022.05.07..
- */
-object ErrorMessage {
-    const val NO_RACING_CARS = "레이싱카가 없습니다."
-    const val ATTEMPTS_IS_ZERO = "시도할 횟수가 0회입니다."
-}
