--- src/main/kotlin/racingCar/etc/PrintingMessage.kt
@@ -1,11 +1,0 @@
-package racingCar.etc
-
-/**
- * 일반적으로 출력되는 메세지를 모아둔 클래스
- * Created by Jaesungchi on 2022.05.07..
- */
-object PrintingMessage {
-    const val HOW_MANY_CAR = "자동차 대수는 몇 대인가요?"
-    const val HOW_MANY_TIMES_RACE = "시도할 횟수는 몇 회인가요?"
-    const val EXECUTION_RESULT = "실행 결과"
-}
